aman 3

+++ routes/web.php
@@ -1,6 +1,8 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+
+
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\PortoController;
 use App\Http\Controllers\GaleriController;
@@ -36,33 +38,32 @@ Route::prefix('admin')->group(function () {
     Route::prefix('artikel')->group(function () {
         Route::get('/', [ArtikelController::class, 'index'])->name('admin.artikel.index');
         Route::get('/create', [ArtikelController::class, 'create'])->name('admin.artikel.create');
-        Route::post('/', [ArtikelController::class, 'store']);
-        Route::get('/{id}', [ArtikelController::class, 'show']);
-        Route::get('/{id}/edit', [ArtikelController::class, 'edit']);
-        Route::put('/{id}', [ArtikelController::class, 'update']);
-        Route::delete('/{id}', [ArtikelController::class, 'destroy']);
+        Route::post('/', [ArtikelController::class, 'store'])->name('admin.artikel.store');
+        Route::get('/edit/{id}', [ArtikelController::class, 'edit'])->name('admin.artikel.edit');
+        Route::put('/{id}', [ArtikelController::class, 'update'])->name('admin.artikel.update');
+        Route::delete('/{id}', [ArtikelController::class, 'destroy'])->name('admin.artikel.destroy');
     });
     Route::prefix('galeri')->group(function () {
         Route::get('/', [GaleriController::class, 'index'])->name('admin.galeri.index');
         Route::get('/create', [GaleriController::class, 'create'])->name('admin.galeri.create');
-        Route::post('/', [GaleriController::class, 'store']);
-        Route::get('/{id}', [GaleriController::class, 'show']);
-        Route::get('/{id}/edit', [GaleriController::class, 'edit']);
-        Route::put('/{id}', [GaleriController::class, 'update']);
-        Route::delete('/{id}', [GaleriController::class, 'destroy']);
+        Route::post('/', [GaleriController::class, 'store'])->name('admin.galeri.store');
+        Route::get('/edit/{id}', [GaleriController::class, 'edit'])->name('admin.galeri.edit');
+        Route::put('/{id}', [GaleriController::class, 'update'])->name('admin.galeri.update');
+        Route::delete('/{id}', [GaleriController::class, 'destroy'])->name('admin.galeri.destroy');
     });
     Route::prefix('portofolio')->group(function () {
         Route::get('/', [PortoController::class, 'index'])->name('admin.porto.index');
         Route::get('/create', [PortoController::class, 'create'])->name('admin.porto.create');
-        Route::post('/', [PortoController::class, 'store']);
-        Route::get('/{id}', [PortoController::class, 'show']);
-        Route::get('/{id}/edit', [PortoController::class, 'edit']);
-        Route::put('/{id}', [PortoController::class, 'update']);
-        Route::delete('/{id}', [PortoController::class, 'destroy']);
+        Route::post('/', [PortoController::class, 'store'])->name('admin.porto.store');
+        Route::get('/edit/{id}', [PortoController::class, 'edit'])->name('admin.porto.edit');
+        Route::put('/{id}', [PortoController::class, 'update'])->name('admin.porto.update');
+        Route::delete('/{id}', [PortoController::class, 'destroy'])->name('admin.porto.destroy');
     });
 });
 
 // login
 Route::get('/login', [LoginController::class, 'index'])->name('login');
-Route::post('/login', [LoginController::class, 'authenticate']);
+Route::post('/postlogin', [LoginController::class, 'authenticate'])->name('login.authenticate');
 Route::get('/logout', [LoginController::class, 'logout'])->name('logout');
+Route::get('/register', [LoginController::class, 'register'])->name('register');
+Route::post('/register', [LoginController::class, 'store'])->name('register.store');
